// ehif-dbi/src/main/frontend/src/app/api/models/read-test-output-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ReadTestOutputDto {
  mongoFindAllTimeMs?: number;
  mongoWithFilterAndProjectionAndSortTimeMs?: number;
  mongoWithFilterAndProjectionTimeMs?: number;
  mongoWithFilterTimeMs?: number;
  postgresFindAllTimeMs?: number;
  postgresWithFilterAndProjectionAndSortTimeMs?: number;
  postgresWithFilterAndProjectionTimeMs?: number;
  postgresWithFilterTimeMs?: number;
}